Add unit tests for class Money
This commit will add a set of unit tests for class Money.

Note that the unit tests carefully avoid testing rounding behavior at
the midpoint.  For example, there is no test to see if 2.555 is
rounded to 2.56 or 2.55.

The reason is that there is ambiguity here, between different
currencies.  The Euro is rounded up at the mid-point[1], and it is
probably the same for the Swedish krona[2].  The Japanese Yen is
always rounded down and there are probably other schemes, like
rounding to the nearest even value (bankers' rounding), for example.

Accounting systems typically allow you to specify rounding rules for
each currency.

Why not try adding a few tests with rounding for a select few currency
types?  What code changes do you need to apply to class Money in order
to correctly round for both Euro and Yen?  Can you better see the
problems with representing currency with a single floating point value
now?

[1] http://ec.europa.eu/economy_finance/publications/pages/publication1224_en.pdf
[2] https://www4.skatteverket.se/rattsligvagledning/edition/2014.3/321578.html