Simplify Linq expression for adding basket total
This commit will add a few extension methods to IEnumerable so that we
can simplify the Linq expression when calculating a Money aggregate.

This means that calculating the total cost of the basket can be done
like this:

    public Money Total => items.Sum(item => item.Cost);

which is exactly the same kind of expression that we did when
representing cost with a double.  This makes the conceptual changes to
the code between using a double and Money more clear, I think.