Host the domain functionality as a REST API
This commit will host the Commerce domain as a REST API, using
ASP.NET.

This includes the following changes to class Money:

1. Explicitally set midpoint rounding algorithm to round toward the
   nearest even number.

2. Support a "none" value, representing zero money without currency
   code.

This has made class Money considerably more complex, but also more
usable.  The None value in particular is very helpful when dealing
with an empty basket, or when writing unit tests.

Also, instead of classes Apple and Banana, there is now a proper
Product domain class.  The code also assumes that there is an external
product REST API that returns products that can be used by our REST
API.