Use settings file for URL configuration in L2 tests
This commit will add a new class, ConfigurableHttpClient, that will
read the base address from a testsettings.json file.  This is used to
allow the user to configure the URL of the hosted REST API without
editing the source code, and messing up their Git status.

The default URL for the REST API is now http://localhost:5000