Refactor settings for database options in L1 tests
This commit will split out the code used to read the testsettings.json
file into a separate class, ConfigurableDatabaseOptions.  It now also
harmonize with the details of class ConfigurableHttpClient, introduced
in commit 45eb0cf80d0.

The format of the testsettings.json class is also updated to avoid
conflicts with standard environment variables such as USERNAME.